# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/requests": {
      "get": {
        "description": "Returns all org registration, budget increase, and other requests",
        "operationId": "RequestsController_findAll",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requests retrieved successfully"
          },
          "403": {
            "description": "Access denied — requires admin or superuser role"
          }
        },
        "summary": "Get all platform requests",
        "tags": [
          "Requests"
        ]
      },
      "post": {
        "operationId": "RequestsController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Request created successfully"
          },
          "400": {
            "description": "Validation failed"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "Create a new platform request",
        "tags": [
          "Requests"
        ]
      }
    },
    "/requests/pending": {
      "get": {
        "operationId": "RequestsController_getPending",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending requests retrieved"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "Get pending requests",
        "tags": [
          "Requests"
        ]
      }
    },
    "/requests/{id}": {
      "get": {
        "operationId": "RequestsController_findOne",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request retrieved"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Request not found"
          }
        },
        "summary": "Get request by ID",
        "tags": [
          "Requests"
        ]
      },
      "patch": {
        "operationId": "RequestsController_update",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request updated"
          },
          "403": {
            "description": "Access denied — requires superuser role"
          },
          "404": {
            "description": "Request not found"
          }
        },
        "summary": "Update a request",
        "tags": [
          "Requests"
        ]
      },
      "delete": {
        "operationId": "RequestsController_remove",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request deleted"
          },
          "403": {
            "description": "Access denied — requires superuser role"
          },
          "404": {
            "description": "Request not found"
          }
        },
        "summary": "Delete a request",
        "tags": [
          "Requests"
        ]
      }
    },
    "/requests/{id}/approve": {
      "patch": {
        "operationId": "RequestsController_approve",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request approved"
          },
          "403": {
            "description": "Access denied — requires superuser role"
          },
          "404": {
            "description": "Request not found"
          }
        },
        "summary": "Approve a request",
        "tags": [
          "Requests"
        ]
      }
    },
    "/requests/{id}/reject": {
      "patch": {
        "operationId": "RequestsController_reject",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role for RBAC",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "volunteer",
                "donor",
                "beneficiary"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request rejected"
          },
          "403": {
            "description": "Access denied — requires superuser role"
          },
          "404": {
            "description": "Request not found"
          }
        },
        "summary": "Reject a request",
        "tags": [
          "Requests"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kindred Platform API",
    "description": "Backend REST API for Kindred — India's Volunteer Platform. Supports Users, Organizations, Volunteers, Programs, Requests, Tasks, Hour Logs, Resource Donations, Program Assignments, Program Applications, and Subtasks. Role-based access control via the \"role\" header.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateRequestDto": {
        "type": "object",
        "properties": {
          "org": {
            "type": "string",
            "description": "Organization name",
            "example": "PathFinders NGO"
          },
          "type": {
            "type": "string",
            "description": "Type of request",
            "example": "Organization Registration"
          },
          "description": {
            "type": "string",
            "description": "Detailed description",
            "example": "New NGO registration for youth empowerment programs in rural Tamil Nadu"
          },
          "priority": {
            "type": "string",
            "description": "Priority level",
            "example": "high"
          },
          "contact": {
            "type": "string",
            "description": "Contact email",
            "example": "admin@pathfinders.org"
          },
          "phone": {
            "type": "string",
            "description": "Contact phone",
            "example": "9876543210"
          },
          "city": {
            "type": "string",
            "description": "City",
            "example": "Chennai"
          }
        },
        "required": [
          "org",
          "type",
          "description"
        ]
      },
      "UpdateRequestDto": {
        "type": "object",
        "properties": {
          "org": {
            "type": "string",
            "description": "Organization name",
            "example": "PathFinders NGO"
          },
          "type": {
            "type": "string",
            "description": "Type of request",
            "example": "Organization Registration"
          },
          "description": {
            "type": "string",
            "description": "Detailed description",
            "example": "New NGO registration for youth empowerment programs in rural Tamil Nadu"
          },
          "priority": {
            "type": "string",
            "description": "Priority level",
            "example": "high"
          },
          "contact": {
            "type": "string",
            "description": "Contact email",
            "example": "admin@pathfinders.org"
          },
          "phone": {
            "type": "string",
            "description": "Contact phone",
            "example": "9876543210"
          },
          "city": {
            "type": "string",
            "description": "City",
            "example": "Chennai"
          }
        }
      },
...
    }
  }
}